[8주차] 기능 구현

+++ wekiki/src/main/java/com/kakao/techcampus/wekiki/group/service/GroupService.java
@@ -1,26 +1,24 @@
-package com.kakao.techcampus.wekiki.group;
+package com.kakao.techcampus.wekiki.group.service;
 
 import com.kakao.techcampus.wekiki._core.error.exception.Exception400;
 import com.kakao.techcampus.wekiki._core.error.exception.Exception404;
 import com.kakao.techcampus.wekiki._core.error.exception.Exception500;
-import com.kakao.techcampus.wekiki.group.groupDTO.requestDTO.CreateUnOfficialGroupRequestDTO;
-import com.kakao.techcampus.wekiki.group.groupDTO.requestDTO.JoinGroupRequestDTO;
-import com.kakao.techcampus.wekiki.group.groupDTO.requestDTO.UpdateMyGroupPageDTO;
-import com.kakao.techcampus.wekiki.group.groupDTO.responseDTO.*;
-import com.kakao.techcampus.wekiki.group.invitation.Invitation;
-import com.kakao.techcampus.wekiki.group.member.ActiveGroupMember;
-import com.kakao.techcampus.wekiki.group.member.GroupMemberJPARepository;
-import com.kakao.techcampus.wekiki.group.member.InactiveGroupMember;
-import com.kakao.techcampus.wekiki.group.officialGroup.OfficialGroup;
-import com.kakao.techcampus.wekiki.group.unOfficialGroup.closedGroup.UnOfficialClosedGroup;
-import com.kakao.techcampus.wekiki.group.unOfficialGroup.openedGroup.UnOfficialOpenedGroup;
+import com.kakao.techcampus.wekiki.group.domain.Group;
+import com.kakao.techcampus.wekiki.group.dto.GroupRequestDTO;
+import com.kakao.techcampus.wekiki.group.dto.GroupResponseDTO;
+import com.kakao.techcampus.wekiki.group.repository.GroupJPARepository;
+import com.kakao.techcampus.wekiki.group.domain.member.ActiveGroupMember;
+import com.kakao.techcampus.wekiki.group.repository.GroupMemberJPARepository;
+import com.kakao.techcampus.wekiki.group.domain.member.InactiveGroupMember;
+import com.kakao.techcampus.wekiki.group.domain.OfficialGroup;
+import com.kakao.techcampus.wekiki.group.domain.UnOfficialClosedGroup;
+import com.kakao.techcampus.wekiki.group.domain.UnOfficialOpenedGroup;
 import com.kakao.techcampus.wekiki.history.History;
 import com.kakao.techcampus.wekiki.history.HistoryJPARepository;
 import com.kakao.techcampus.wekiki.member.Member;
 import com.kakao.techcampus.wekiki.member.MemberJPARepository;
 import com.kakao.techcampus.wekiki.post.Post;
 import com.kakao.techcampus.wekiki.post.PostJPARepository;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -46,7 +44,7 @@ public class GroupService {
         비공식 그룹 생성
      */
     @Transactional
-    public CreateUnOfficialGroupResponseDTO createUnOfficialGroup(CreateUnOfficialGroupRequestDTO requestDTO, Long memberId) {
+    public GroupResponseDTO.CreateUnOfficialGroupResponseDTO createUnOfficialGroup(GroupRequestDTO.CreateUnOfficialGroupRequestDTO requestDTO, Long memberId) {
         try {
             // Group 생성
             Group group = switch (requestDTO.groupType()) {
@@ -65,13 +63,11 @@ public class GroupService {
             groupMemberJPARepository.save(groupMember);
 
             // return
-            return new CreateUnOfficialGroupResponseDTO(group);
+            return new GroupResponseDTO.CreateUnOfficialGroupResponseDTO(group);
 
-        } catch (Exception400 e) {
+        } catch (Exception400 | Exception404 e) {
             throw e;
-        } catch (Exception404 e) {
-            throw e;
-        }   catch (Exception e) {
+        } catch (Exception e) {
             throw new Exception500("서버 에러가 발생했습니다.");
         }
     }
@@ -80,12 +76,11 @@ public class GroupService {
     /*
         비공식 비공개 그룹 생성 후 반환
      */
-    protected UnOfficialClosedGroup buildUnOfficialClosedGroup(CreateUnOfficialGroupRequestDTO requestDTO) {
+    protected UnOfficialClosedGroup buildUnOfficialClosedGroup(GroupRequestDTO.CreateUnOfficialGroupRequestDTO requestDTO) {
         // Group 생성
         return UnOfficialClosedGroup.unOfficialClosedGroupBuilder()
                 .groupName(requestDTO.groupName())
                 .groupProfileImage(requestDTO.groupImage())
-                .memberCount(1)
                 .created_at(LocalDateTime.now())
                 .build();
     }
@@ -93,12 +88,11 @@ public class GroupService {
     /*
         비공식 공개 그룹 생성 후 반환
      */
-    protected UnOfficialOpenedGroup buildUnOfficialOpenedGroup(CreateUnOfficialGroupRequestDTO requestDTO) {
+    protected UnOfficialOpenedGroup buildUnOfficialOpenedGroup(GroupRequestDTO.CreateUnOfficialGroupRequestDTO requestDTO) {
         // Group 생성
         return UnOfficialOpenedGroup.unOfficialOpenedGroupBuilder()
                 .groupName(requestDTO.groupName())
                 .groupProfileImage(requestDTO.groupImage())
-                .memberCount(1)
                 .created_at(LocalDateTime.now())
                 .introduction(requestDTO.introduction())
                 .entranceHint(requestDTO.entranceHint())
@@ -122,7 +116,7 @@ public class GroupService {
         공식 그룹 리스트, 비공식 공개 그룹 리스트 그룹
         - keyword가 포함된 모든 그룹 그룹별 리스트
      */
-    public SearchGroupDTO searchGroupByKeyword(String keyword) {
+    public GroupResponseDTO.SearchGroupDTO searchGroupByKeyword(String keyword) {
         try {
             Pageable pageable = PageRequest.of(0, 10);
 
@@ -131,7 +125,7 @@ public class GroupService {
             // 비공식 공개 그룹 리스트
             Page<UnOfficialOpenedGroup> unOfficialOpenedGroups = groupJPARepository.findUnOfficialOpenedGroupsByKeyword(keyword, pageable);
 
-            return new SearchGroupDTO(officialGroups, unOfficialOpenedGroups);
+            return new GroupResponseDTO.SearchGroupDTO(officialGroups, unOfficialOpenedGroups);
 
         } catch (Exception e) {
             throw new Exception500("서버 에러가 발생했습니다.");
@@ -141,14 +135,14 @@ public class GroupService {
     /*
         공식 그룹 추가 리스트
      */
-    public SearchOfficialGroupResponseDTO searchOfficialGroupByKeyword(String keyword, int page, int size) {
+    public GroupResponseDTO.SearchOfficialGroupResponseDTO searchOfficialGroupByKeyword(String keyword, int page, int size) {
         try {
             Pageable pageable = PageRequest.of(page, size);
 
             // 비공식 공개 그룹 리스트
             Page<OfficialGroup> officialGroups = groupJPARepository.findOfficialGroupsByKeyword(keyword, pageable);
 
-            return new SearchOfficialGroupResponseDTO(officialGroups);
+            return new GroupResponseDTO.SearchOfficialGroupResponseDTO(officialGroups);
 
         } catch (Exception e) {
             throw new Exception500("서버 에러가 발생했습니다.");
@@ -158,73 +152,31 @@ public class GroupService {
     /*
         비공식 공개 그룹 추가 리스트
      */
-    public SearchUnOfficialGroupResponseDTO searchUnOfficialGroupByKeyword(String keyword, int page, int size) {
+    public GroupResponseDTO.SearchUnOfficialGroupResponseDTO searchUnOfficialGroupByKeyword(String keyword, int page, int size) {
         try {
             Pageable pageable = PageRequest.of(page, size);
 
             // 비공식 공개 그룹 리스트
             Page<UnOfficialOpenedGroup> unOfficialOpenedGroups = groupJPARepository.findUnOfficialOpenedGroupsByKeyword(keyword, pageable);
 
-            return new SearchUnOfficialGroupResponseDTO(unOfficialOpenedGroups);
+            return new GroupResponseDTO.SearchUnOfficialGroupResponseDTO(unOfficialOpenedGroups);
 
         } catch (Exception e) {
             throw new Exception500("서버 에러가 발생했습니다.");
         }
     }
-
-    /*
-        초대 링크 확인
-     */
-    public GetInvitationLinkResponseDTO getGroupInvitationLink(Long groupId) {
-        try {
-            Invitation invitation = groupJPARepository.findInvitationLinkByGroupId(groupId);
-
-            return new GetInvitationLinkResponseDTO(invitation);
-
-        } catch (Exception e) {
-            throw new Exception500("서버 에러가 발생했습니다.");
-        }
-    }
-
-    /*
-        비공식 비공개 그룹 초대 링크 확인
-        - 확인 후 해당하는 그룹 Id 반환
-        - 가입 url 생성 후 그룹 가입 API
-     */
-    public ValidateInvitationResponseDTO ValidateInvitation(String invitationLink) {
-        try {
-            String[] invitation = invitationLink.split("/");
-
-            Long groupId = Long.parseLong(invitation[0]);
-            String invitationCode = invitation[1];
-
-            UnOfficialClosedGroup group = groupJPARepository.findUnOfficialClosedGroupById(groupId);
-
-            // 초대 코드가 틀린 경우
-            if (!group.getInvitation().getInvitationCode().equals(invitationCode)) {
-                throw new Exception400("초대 코드가 일치하지 않습니다.");
-            }
-
-            return new ValidateInvitationResponseDTO(groupId);
-
-        } catch (Exception400 e) {
-            throw e;
-        } catch (Exception e) {
-            throw new Exception500("서버 에러가 발생했습니다.");
-        }
-    }
-
 
     /*
         비공식 공개 그룹 상세 정보 조회
      */
-    public SearchGroupInfoDTO getGroupInfo(Long groupId) {
+    public GroupResponseDTO.SearchGroupInfoDTO getGroupInfo(Long groupId) {
         try {
-            UnOfficialOpenedGroup group = groupJPARepository.findUnOfficialOpenedGroupById(groupId);
-            return new SearchGroupInfoDTO(group);
+            UnOfficialOpenedGroup group = groupJPARepository.findUnOfficialOpenedGroupById(groupId)
+                    .orElseThrow(() -> new Exception404("그룹을 찾을 수 없습니다."));
 
-        } catch (EntityNotFoundException e) {
-            throw new Exception404("그룹을 찾을 수 없습니다.");
+            return new GroupResponseDTO.SearchGroupInfoDTO(group);
+        } catch (Exception404 e) {
+            throw e;
         } catch (Exception e) {
             throw new Exception500("서버 에러가 발생했습니다.");
         }
@@ -234,19 +186,18 @@ public class GroupService {
     /*
         비공식 공개 그룹 입장
      */
-    public void groupEntry(Long groupId, String entrancePassword) {
+    public void groupEntry(Long groupId, GroupRequestDTO.GroupEntryRequestDTO requestDTO) {
         try {
-            UnOfficialOpenedGroup group = groupJPARepository.findUnOfficialOpenedGroupById(groupId);
+            UnOfficialOpenedGroup group = groupJPARepository.findUnOfficialOpenedGroupById(groupId)
+                    .orElseThrow(() -> new Exception404("그룹을 찾을 수 없습니다."));
 
             // 틀린 경우, 에러 핸들링
-            if(!group.getEntrancePassword().equals(entrancePassword)) {
+            if(!group.getEntrancePassword().equals(requestDTO.entrancePassword())) {
                 throw new Exception400("비밀번호가 틀렸습니다.");
             }
 
-        } catch (Exception400 e) {
+        } catch (Exception400 | Exception404 e) {
             throw e;
-        }  catch (EntityNotFoundException e) {
-            throw new Exception404("그룹을 찾을 수 없습니다.");
         } catch (Exception e) {
             throw new Exception500("서버 에러가 발생했습니다.");
         }
@@ -256,28 +207,41 @@ public class GroupService {
         그룹 참가 (공통 부분)
      */
     @Transactional
-    public void joinGroup(Long groupId, Long memberId, JoinGroupRequestDTO requestDTO) {
-        try {
-            // 회원 정보 확인
-            Member member = getMemberById(memberId);
+    public void joinGroup(Long groupId, Long memberId, GroupRequestDTO.JoinGroupRequestDTO requestDTO) {
+        // 회원 정보 확인
+        Member member = getMemberById(memberId);
 
-            // 그룹 정보 확인
-            // TODO: Redis 활용
+        // 그룹 정보 확인
+        Group group = getGroupById(groupId);
+
+        // 이미 가입한 상태일 시 예외 처리
+        if (groupMemberJPARepository.findActiveGroupMemberByMemberIdAndGroupId(memberId, groupId).isPresent()) {
+            throw new Exception400("이미 가입된 회원입니다.");
+        }
+
+        // 재가입 회원인지 확인
+        InactiveGroupMember wasGroupMember = groupMemberJPARepository.findInactiveGroupMemberByMemberAndGroup(member, group);
+
+        // 재가입 회원이면 활성화, 신규 회원이면 새로 생성
+        ActiveGroupMember groupMember = wasGroupMember != null ? new ActiveGroupMember(wasGroupMember) : buildGroupMember(member, group, requestDTO.nickName());
+
+        // 그룹 멤버 잔재 삭제
+        if(wasGroupMember != null) {
+            groupMemberJPARepository.delete(wasGroupMember);
+        }
+
+        // GroupMember 저장
+        groupMemberJPARepository.save(groupMember);
+    }
+
+    /*
+        그룹 내 그룹원 리스트 조회
+     */
+    public GroupResponseDTO.GetGroupMembersResponseDTO getGroupMembers(Long groupId) {
+        try {
             Group group = getGroupById(groupId);
 
-            // 재가입 회원인지 확인
-            InactiveGroupMember wasGroupMember = groupMemberJPARepository.findInactiveGroupMemberByMemberAndGroup(member, group);
-
-            // 재가입 회원이면 활성화, 신규 회원이면 새로 생성
-            ActiveGroupMember groupMember = wasGroupMember != null ? new ActiveGroupMember(wasGroupMember) : buildGroupMember(member, group, requestDTO.nickName());
-
-            // 그룹 멤버 잔재 삭제
-            if(wasGroupMember != null) {
-                groupMemberJPARepository.delete(wasGroupMember);
-            }
-
-            // GroupMember 저장
-            groupMemberJPARepository.save(groupMember);
+            return new GroupResponseDTO.GetGroupMembersResponseDTO(group);
 
         } catch (Exception404 e) {
             throw e;
@@ -289,28 +253,22 @@ public class GroupService {
     /*
         그룹 내 마이 페이지
      */
-    public MyGroupInfoResponseDTO getMyGroupInfo(Long groupId, Long memberId) {
+    public GroupResponseDTO.MyGroupInfoResponseDTO getMyGroupInfo(Long groupId, Long memberId) {
         try {
-            // 회원 정보 확인
-            Member member = getMemberById(memberId);
-
-            // 그룹 정보 확인
-            // TODO: Redis 활용
-            Group group = getGroupById(groupId);
-
             // 그룹 멤버 확인
-            ActiveGroupMember groupMember = groupMemberJPARepository.findActiveGroupMemberByMemberAndGroup(member, group);
+            ActiveGroupMember groupMember = groupMemberJPARepository.findActiveGroupMemberByMemberIdAndGroupId(memberId, groupId)
+                    .orElseThrow(() -> new Exception404("해당 그룹의 회원이 아닙니다"));
 
             // 해당 멤버의 Post 기록 정보 확인(History에서 가져옴)
             Pageable pageable = PageRequest.of(0, 10);
             Page<History> myHistoryList = historyJPARepository.findAllByGroupMember(groupMember, pageable);
 
             // 그룹 이름, 현재 닉네임, Post 기록 정보를 담은 responseDTO 반환
-            return new MyGroupInfoResponseDTO(group, groupMember, myHistoryList);
+            return new GroupResponseDTO.MyGroupInfoResponseDTO(groupMember.getGroup(), groupMember, myHistoryList);
 
         } catch (Exception404 e) {
             throw e;
-        }  catch (Exception e) {
+        } catch (Exception e) {
             throw new Exception500("서버 에러가 발생했습니다.");
         }
     }
@@ -319,22 +277,16 @@ public class GroupService {
         내 문서 기여 목록 전체 보기
      */
     @Transactional
-    public MyGroupHistoryResponseDTO getMyGroupHistory(Long groupId, Long memberId, int page, int size) {
+    public GroupResponseDTO.MyGroupHistoryResponseDTO getMyGroupHistory(Long groupId, Long memberId, int page, int size) {
         try {
-            // 회원 정보 확인
-            Member member = getMemberById(memberId);
-
-            // 그룹 정보 확인
-            // TODO: Redis 활용
-            Group group = getGroupById(groupId);
-
             // 그룹 멤버 확인
-            ActiveGroupMember groupMember = groupMemberJPARepository.findActiveGroupMemberByMemberAndGroup(member, group);
+            ActiveGroupMember groupMember = groupMemberJPARepository.findActiveGroupMemberByMemberIdAndGroupId(memberId, groupId)
+                    .orElseThrow(() -> new Exception404("해당 그룹의 회원이 아닙니다"));
 
             Pageable pageable = PageRequest.of(page, size);
             Page<History> myHistoryList = historyJPARepository.findAllByGroupMember(groupMember, pageable);
 
-            return new MyGroupHistoryResponseDTO(myHistoryList);
+            return new GroupResponseDTO.MyGroupHistoryResponseDTO(myHistoryList);
 
         } catch (Exception404 e) {
             throw e;
@@ -348,18 +300,11 @@ public class GroupService {
         그룹 내 마이 페이지 정보 수정
      */
     @Transactional
-    public void updateMyGroupPage(Long groupId, Long memberId, UpdateMyGroupPageDTO requestDTO) {
+    public void updateMyGroupPage(Long groupId, Long memberId, GroupRequestDTO.UpdateMyGroupPageDTO requestDTO) {
         try {
-            // 회원 정보 확인
-            Member member = getMemberById(memberId);
-
-            // 그룹 정보 확인
-            // TODO: Redis 활용
-            Group group = getGroupById(groupId);
-
             // 그룹 멤버 확인
-            // TODO: Redis 활용
-            ActiveGroupMember groupMember = groupMemberJPARepository.findActiveGroupMemberByMemberAndGroup(member, group);
+            ActiveGroupMember groupMember = groupMemberJPARepository.findActiveGroupMemberByMemberIdAndGroupId(memberId, groupId)
+                    .orElseThrow(() -> new Exception404("해당 그룹의 회원이 아닙니다"));
 
             // 그룹 닉네임 변경
             groupMember.update(requestDTO.groupNickName());
@@ -381,16 +326,9 @@ public class GroupService {
     @Transactional
     public void leaveGroup(Long groupId, Long memberId) {
         try {
-            // 회원 정보 확인
-            Member member = getMemberById(memberId);
-
-            // 그룹 정보 확인
-            // TODO: Redis 활용
-            Group group = getGroupById(groupId);
-
             // 그룹 멤버 확인
-            // TODO: Redis 활용
-            ActiveGroupMember activeGroupMember = groupMemberJPARepository.findActiveGroupMemberByMemberAndGroup(member, group);
+            ActiveGroupMember activeGroupMember = groupMemberJPARepository.findActiveGroupMemberByMemberIdAndGroupId(memberId, groupId)
+                    .orElseThrow(() -> new Exception404("해당 그룹의 회원이 아닙니다"));
 
             // 탈퇴 그룹 회원 생성
             InactiveGroupMember inactiveGroupMember = new InactiveGroupMember(activeGroupMember);
@@ -422,39 +360,4 @@ public class GroupService {
     protected Group getGroupById(Long groupId) {
         return groupJPARepository.findById(groupId).orElseThrow(() -> new Exception404("해당 그룹을 찾을 수 없습니다."));
     }
-
-    // RedisCache를 통한 getMember, getGroup -> 근데 성능 개선에 도움이 될까?
-    /*
-        protected Member getMemberById(Long memberId) {
-            try {
-                String memberKey = "member:" + memberId;
-                ValueOperations<String, Object> operations = redisTemplate.opsForValue();
-                if(redisTemplate.hasKey(memberKey)) {
-                    return (Member) operations.get(memberKey);
-                } else {
-                    Member member = memberJPARepository.findById(memberId).orElseThrow(() -> new Exception404("해당 사용자를 찾을 수 없습니다."));
-                    operations.set(memberKey, member);
-                    return member;
-                }
-            } catch (Exception e) {
-                throw new Exception500("서버 에러가 발생했습니다.");
-            }
-        }
-
-        protected Group getGroupById(Long groupId) {
-            try {
-                String groupKey = "group:" + groupId;
-                ValueOperations<String, Object> operations = redisTemplate.opsForValue();
-                if(redisTemplate.hasKey(groupKey)) {
-                    return (Group) operations.get(groupKey);
-                } else {
-                    Group group = groupJPARepository.findById(groupId).orElseThrow(() -> new Exception404("해당 그룹을 찾을 수 없습니다."));
-                    operations.set(groupKey, group);
-                    return group;
-                }
-            } catch (Exception e) {
-                throw new Exception500("서버 에러가 발생했습니다.");
-            }
-        }
-     */
 }
